Views-structure
-Split views for views.py file
