+++ app/views/__init__.py
@@ -1,0 +1,3 @@
+from .index import index
+from .dashboard import *
+from .formulario import *
