+++ project/urls.py
@@ -20,6 +20,8 @@ from app import views
 from django.conf.urls.static import static
 from django.conf import settings
 
+
+
 urlpatterns = [
     path('',views.index, name='index'),
     path('admin/', admin.site.urls),
